lab_6 + start script
